+++ src/daemon/main.c
@@ -483,6 +483,11 @@ main(int argc, char **argv)
     /* Release PID file lock */
     release_pidfile_lock();
 
+    /* Free command-line allocated strings (fixes ASan-detected leak) */
+    g_free((gchar*)conffile);
+    g_free((gchar*)statefile);
+    g_free((gchar*)logfile);
+
     g_debug("exiting");
     return EXIT_SUCCESS;
 }